requirements.txt
Django>=4.2.20,<4.3
django-compressor>=4.5.1,<4.6
django-crispy-forms>=2.4,<2.5
crispy-bootstrap3>=2024.1,<2024.2
django-debug-toolbar>=5.1.0,<5.2
django-wkhtmltopdf>=3.4.0,<3.5
mysqlclient>=2.1.1
icalendar==4.1.0
jsonfield>=3.1.0,<3.2
# Django Modern RPC -- For our JSONRPC API
django-modern-rpc>=1.0.1,<1.1
raven==6.10.0
# Translations -- development
django-rosetta>=0.10.2,<0.11.0
# Django-extensions (runserver_plus command) -- development
django-extensions>=4.1,<4.2
# Werkzeug needed for runserver_plus command) -- development
Werkzeug>=3.1,<3.2
# Single Sign On - OIDC Client
mozilla-django-oidc>=4.0.1,<4.1
# Sentry error logging
sentry-sdk>=2.26.0,<2.27
# UWSGI - to run the website in WSGI mode on a proper webserver
uwsgi>=2.0.29,<2.1
# Configure Django via environment variables
django-environ>=0.12.0,<0.13
